+++ app/__main__.py
@@ -1,9 +1,16 @@
 #!/usr/bin/env python
 """Redirects to entrypoint of the app"""
 
+from os import environ as env
 from sys import exit
+
+# TODO conditional imports and env best practices
+# conditional to avoid flake8 E402 module level import not at top of file
+if True:
+    env["APP_DEBUG_IS_ON"] = "True"
 
 from .app import main
 
 if __name__ == "__main__":
-    exit(main(debug_on=True))
+    # main(sys.argv[1], sys.argv[2], sys.argv[3])
+    exit(main())