+++ app/helper/load_configs.py
@@ -1,21 +1,14 @@
 #!/usr/bin/env python
-"""
-Load configuration files
-TODO some error checks missing
-"""
+"""Load configuration files"""
 
-from logging import debug  # , error
+from os import environ as env
 from os.path import join  # , expanduser , abspath, dirname, exists, split
 from typing import Literal, Optional, Union
 
+if "APP_DEBUG_IS_ON" in env:
+    from logging import debug
+
 from omegaconf import OmegaConf
-
-# from yaml import safe_load
-
-
-def set_debug_state_cfg(debug_on: bool = False):
-    global debug_on_glob
-    debug_on_glob = debug_on
 
 
 def load_defaults(
@@ -23,34 +16,33 @@ def load_defaults(
 ) -> None:
     global default_configs
     default_configs = _load_config(cfg_defaults_fn, cfg_path)
-    if debug_on_glob:
+    if "APP_DEBUG_IS_ON" in env:
         debug(f"{default_configs=}")
 
 
 def get_config_content(
-    cfg_filename: str = "defaults", cfg_path: str = "app/config"
+    cfg_filename_ex_ext: str = "defaults", cfg_path: str = "app/config"
 ) -> dict:
     """Returns config objects"""
 
-    if debug_on_glob:
-        defaults_dbg_msg = f"'{cfg_filename}' found in defaults."
-        # TODO test case and type
-        if _check_or_get_default("config_fn", cfg_filename):
+    if "APP_DEBUG_IS_ON" in env:
+        defaults_dbg_msg = f"'{cfg_filename_ex_ext}' found in defaults."
+        if _check_or_get_default("config_fn", cfg_filename_ex_ext):
             debug(defaults_dbg_msg)
         else:
             debug(f"No {defaults_dbg_msg} Trying to find the file anyway.")
 
-    return _load_config(cfg_filename)
+    return _load_config(cfg_filename_ex_ext, cfg_path)
 
 
 def _load_config(
     cfg_filename_ex_ext: str = "defaults", cfg_path: str = "app/config"
 ) -> OmegaConf:
     """Loads and returns a config. Only accepts yaml/yml."""
-    # TODO sanitization of yaml/yml extension
-    if debug_on_glob:
+    if "APP_DEBUG_IS_ON" in env:
         debug(f"Loading {cfg_filename_ex_ext=}")
     try:
+        # TODO sanitization of yaml/yml extension
         config = OmegaConf.load(join(cfg_path, f"{cfg_filename_ex_ext}.yml"))
     except Exception as e:
         return e
@@ -120,7 +112,7 @@ def _check_or_get_default(
 #     cfg_abs_path = split(dirname(abspath(__file__)))[0]
 #     cfg_abs_path = join(cfg_abs_path, cfg_path, f"{cfg_filename}.yml")
 
-#     if debug_on_glob:
+#     if environ["APP_DEBUG_IS_ON"]:
 #         debug(f"load: {cfg_abs_path}")
 
 #     if not exists(cfg_abs_path):