+++ app/helper/parse_configs_into_paramdict.py
@@ -6,9 +6,11 @@
 
 # from logging import debug
 from json import dump
-from logging import debug
-from os import environ
+from os import environ as env
 from typing import Union
+
+if "APP_DEBUG_IS_ON" in env:
+    from logging import debug
 
 from torch import device
 from torch.cuda import is_available
@@ -18,13 +20,6 @@ from .load_configs import (  # get_default_save_dir,
     get_keyfile_content,
     load_defaults,
 )
-
-# from logging import debug, error, getLogger, root
-
-
-def set_debug_state_parse(debug_on: bool = False):
-    global debug_state
-    debug_state = debug_on
 
 
 def get_param_dict() -> dict:
@@ -69,7 +64,7 @@ def get_param_dict() -> dict:
                 wandb_params, paramobj["project_name"]
             )
 
-    if debug_state:
+    if "APP_DEBUG_IS_ON" in env:
         paramobj_file = "./paramobj.json"
         debug(f"Printing paramobj to '{paramobj_file}'")
         with open(paramobj_file, "w") as outfile:
@@ -130,7 +125,7 @@ def _get_device() -> str:
     """Returns the device as 'cpu', 'gpu' or 'tpu'"""
 
     # TODO make it platform independent, TPU_NAME used by Google Colab
-    if "TPU_NAME" in environ:
+    if "TPU_NAME" in env:
         return "tpu"
     else:
         try: