+++ app/helper/prepare_ml_input.py
@@ -7,8 +7,10 @@ Components could be models, datasets, tokenizers, metrics etc.
 # TODO return None or Exception ?
 # TODO load local versions of models, datasets, metrics and tokenizer if not cached
 # TODO function def with actual objects, not placeholder 'object'
-from logging import debug
-from os import environ
+from os import environ as env
+
+if "APP_DEBUG_IS_ON" in env:
+    from logging import debug
 
 from datasets import dataset_dict
 from transformers import AutoTokenizer
@@ -18,17 +20,26 @@ from .load_hf_components import (
     get_metrics_to_load_objects_hf,
     get_model_hf,
     get_tokenizer_hf,
-    set_debug_state_hf,
 )
 
 
-def set_debug_state_ml(debug_on: bool = False):
-    global debug_state
-    debug_state = debug_on
-    set_debug_state_hf(debug_on)
+def prepare_pipeline(paramobj: dict) -> dict:
+    """TODO"""
+
+    paramobj["dataset"]["num_labels"] = _get_dataset(
+        paramobj["dataset"], paramobj["model_full_name"]
+    )
+    paramobj["metrics"]["metrics_loaded"] = _get_metrics_to_load_objects(
+        paramobj["metrics"]["metrics_to_load"]
+    )
+    _get_model(paramobj["model_full_name"], paramobj["dataset"]["num_labels"])
+    provider = paramobj["sweep"]["provider"]
+    _set_provider_env(provider, paramobj[provider])
+
+    return paramobj
 
 
-def get_dataset(dataset: dict, model_full_name: str) -> int:
+def _get_dataset(dataset: dict, model_full_name: str) -> int:
     """Load tokenized dataset and model"""
 
     try:
@@ -47,7 +58,7 @@ def get_dataset(dataset: dict, model_full_name: str) -> int:
         return e
 
 
-def get_metrics_to_load_objects(metrics_to_load: list) -> list[dict]:
+def _get_metrics_to_load_objects(metrics_to_load: list) -> list[dict]:
     """
     Downloads metrics objects by calling the appropriate handling function.
     To date onyl from Hugging Face
@@ -56,19 +67,19 @@ def get_metrics_to_load_objects(metrics_to_load: list) -> list[dict]:
     return get_metrics_to_load_objects_hf(metrics_to_load)
 
 
-def get_model(model_full_name: str = None, num_labels: str = None):
+def _get_model(model_full_name: str = None, num_labels: str = None):
     """TODO"""
     # TODO implement other providers
     return get_model_hf(model_full_name, num_labels)
 
 
-def set_provider_env(provider: str, provider_param: dict) -> None:
+def _set_provider_env(provider: str, provider_param: dict) -> None:
     """Set the environment parameters for the sweep provider"""
 
     try:
         for k, v in provider_param.items():
-            environ[k] = v
-            if debug_state:
+            env[k] = v
+            if "APP_DEBUG_IS_ON" in env:
                 debug(f"Setting '{provider}' env '{k}'")
     except Exception as e:
         return e
@@ -94,7 +105,7 @@ def _get_tokenized_dataset(
 ) -> dataset_dict.DatasetDict:
     """TODO"""
 
-    if debug_state:
+    if "APP_DEBUG_IS_ON" in env:
         debug(f"Tokenizing dataset with {len(cols_to_tok)} columns to tokenize and")
         debug(f"Removing {cols_to_tok=} and {cols_to_remove=} from tokenized dataset")